refactored how warning level was handled, split data retrieval logic
