+++ src/retrieve-w3c-validation/retrieveForNode.ts
@@ -1,0 +1,22 @@
+// Imports
+import * as https from 'https';
+import { W3CCSSValidatorResponse } from '../types';
+
+// Utility function for retrieving response from W3C CSS Validator in a nNde environment
+const retrieveForNode = async (url: string): Promise<W3CCSSValidatorResponse['cssvalidation']> => {
+	return new Promise((resolve) => {
+		https.get(url, (res) => {
+			let data = '';
+
+			res.on('data', (chunk) => {
+				data += chunk;
+			});
+
+			res.on('end', () => {
+				resolve((JSON.parse(data) as W3CCSSValidatorResponse).cssvalidation);
+			});
+		});
+	});
+};
+
+export default retrieveForNode;
